+++ packages/contracts/src/plugins/governance/majority-voting/token/TokenVotingSetup.sol
@@ -11,7 +11,7 @@ import {IVotesUpgradeable} from "@openzeppelin/contracts-upgradeable/governance/
 import {IDAO} from "../../../../core/dao/IDAO.sol";
 import {DAO} from "../../../../core/dao/DAO.sol";
 import {PermissionLib} from "../../../../core/permission/PermissionLib.sol";
-import {PluginSetup, IPluginSetup} from "../../../../framework/plugin/setup/PluginSetup.sol";
+import {PluginSetupUpgradeable, IPluginSetup} from "../../../../framework/plugin/setup/PluginSetupUpgradeable.sol";
 import {GovernanceERC20} from "../../../../token/ERC20/governance/GovernanceERC20.sol";
 import {GovernanceWrappedERC20} from "../../../../token/ERC20/governance/GovernanceWrappedERC20.sol";
 import {IGovernanceWrappedERC20} from "../../../../token/ERC20/governance/IGovernanceWrappedERC20.sol";
@@ -21,13 +21,10 @@ import {TokenVoting} from "./TokenVoting.sol";
 /// @title TokenVotingSetup
 /// @author Aragon Association - 2022-2023
 /// @notice The setup contract of the `TokenVoting` plugin.
-contract TokenVotingSetup is PluginSetup {
+contract TokenVotingSetup is PluginSetupUpgradeable {
     using Address for address;
     using Clones for address;
     using ERC165Checker for address;
-
-    /// @notice The address of the `TokenVoting` base contract.
-    TokenVoting private immutable tokenVotingBase;
 
     /// @notice The address of the `GovernanceERC20` base contract.
     address public immutable governanceERC20Base;
@@ -63,8 +60,7 @@ contract TokenVotingSetup is PluginSetup {
     constructor(
         GovernanceERC20 _governanceERC20Base,
         GovernanceWrappedERC20 _governanceWrappedERC20Base
-    ) {
-        tokenVotingBase = new TokenVoting();
+    ) PluginSetupUpgradeable(address(new TokenVoting())) {
         governanceERC20Base = address(_governanceERC20Base);
         governanceWrappedERC20Base = address(_governanceWrappedERC20Base);
     }
@@ -137,9 +133,10 @@ contract TokenVotingSetup is PluginSetup {
 
         // Prepare and deploy plugin proxy.
         plugin = createERC1967Proxy(
-            address(tokenVotingBase),
+            implementation,
             abi.encodeWithSelector(TokenVoting.initialize.selector, _dao, votingSettings, token)
         );
+        TokenVoting tokenVoting = TokenVoting(plugin);
 
         // Prepare permissions
         PermissionLib.MultiTargetPermission[]
@@ -154,7 +151,7 @@ contract TokenVotingSetup is PluginSetup {
             plugin,
             _dao,
             PermissionLib.NO_CONDITION,
-            tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
+            tokenVoting.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
         permissions[1] = PermissionLib.MultiTargetPermission(
@@ -162,7 +159,7 @@ contract TokenVotingSetup is PluginSetup {
             plugin,
             _dao,
             PermissionLib.NO_CONDITION,
-            tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
+            tokenVoting.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
         // Grant `EXECUTE_PERMISSION` of the DAO to the plugin.
@@ -201,6 +198,8 @@ contract TokenVotingSetup is PluginSetup {
             revert WrongHelpersArrayLength({length: helperLength});
         }
 
+        TokenVoting tokenVoting = TokenVoting(_payload.plugin);
+
         // token can be either GovernanceERC20, GovernanceWrappedERC20, or IVotesUpgradeable, which
         // does not follow the GovernanceERC20 and GovernanceWrappedERC20 standard.
         address token = _payload.currentHelpers[0];
@@ -217,7 +216,7 @@ contract TokenVotingSetup is PluginSetup {
             _payload.plugin,
             _dao,
             PermissionLib.NO_CONDITION,
-            tokenVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
+            tokenVoting.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
         permissions[1] = PermissionLib.MultiTargetPermission(
@@ -225,7 +224,7 @@ contract TokenVotingSetup is PluginSetup {
             _payload.plugin,
             _dao,
             PermissionLib.NO_CONDITION,
-            tokenVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
+            tokenVoting.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
         permissions[2] = PermissionLib.MultiTargetPermission(
@@ -248,11 +247,6 @@ contract TokenVotingSetup is PluginSetup {
                 GovernanceERC20(token).MINT_PERMISSION_ID()
             );
         }
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
-        return address(tokenVotingBase);
     }
 
     /// @notice Retrieves the interface identifiers supported by the token contract.