+++ packages/contracts/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.3.2-rc0
+
+### Added
+
+- Added an abstract `PluginSetupUpgradeable` base contract.
+
+### Changed
+
+- Added a `address public immutable implementation` variable to `PluginSetup` and `PluginSetupUpgradeable` and removed the `implementation()` functions.
+
+### Removed
+
 ## v1.3.0-rc0
 
 ### Added